backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
importlib-metadata==6.8.0
inflection==0.5.1
kombu==5.3.2
Markdown==3.4.4
mysqlclient==2.2.0
packaging==23.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
zipp==3.17.0
